Dynamically determine histogram view off data / histogram range
Fix breakpoints being hidden
