--- app-frontend/src/app/components/histogram/histogramBreakpoint/histogramBreakpoint.component.js
@@ -1,18 +1,0 @@
-import histogramBreakpointTpl from './histogramBreakpoint.html';
-
-const histogramBreakpoint = {
-    templateUrl: histogramBreakpointTpl,
-    controller: 'HistogramBreakpointController',
-    bindings: {
-        color: '<',
-        breakpoint: '<',
-        range: '<',
-        precision: '<',
-        options: '<',
-        upperBound: '<',
-        lowerBound: '<',
-        onBreakpointChange: '&'
-    }
-};
-
-export default histogramBreakpoint;
